python_liquid2/requirements.txt
babel==2.16.0
iniconfig==2.0.0
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-liquid2==0.1.0
pytz==2024.2
ruff==0.9.1
six==1.17.0
typing_extensions==4.12.2
